+++ jsonargparse/_from_config.py
@@ -8,7 +8,7 @@ from ._common import parser_context
 from ._core import ArgumentParser
 from ._loaders_dumpers import get_loader_exceptions, load_value
 from ._optionals import _get_config_read_mode
-from ._paths import change_to_path_dir
+from ._paths import path_dir_context
 from ._required import clear_required, iter_required_keys
 from ._typehints import is_subclass_spec, resolve_class_path_by_name
 from ._util import import_object, load_config_path_context
@@ -71,7 +71,7 @@ def _parse_class_kwargs_from_config(cls: type[T], config: str | PathLike | dict,
         cfg_path = Path(config, mode=_get_config_read_mode())
         with (
             load_config_path_context(cfg_path),
-            change_to_path_dir(cfg_path),
+            path_dir_context(cfg_path),
             parser_context(load_value_mode=parser.parser_mode),
         ):
             cfg_str = cfg_path.read_text()
@@ -94,7 +94,7 @@ def _parse_class_kwargs_from_config(cls: type[T], config: str | PathLike | dict,
     parser.add_class_arguments(cls)
     for required in iter_required_keys(parser):
         clear_required(parser, required)
-    with load_config_path_context(cfg_path), change_to_path_dir(cfg_path):
+    with load_config_path_context(cfg_path), path_dir_context(cfg_path):
         cfg = parser.parse_object(config, defaults=False)
     return parser.instantiate(cfg).as_dict(), cls